Move SentryHubServiceProvider to Error namespace
