+++ app/Error/SentryHubServiceProvider.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\JwtProvider;
+namespace CultuurNet\UDB3\JwtProvider\Error;
 
+use CultuurNet\UDB3\JwtProvider\BaseServiceProvider;
 use Sentry\SentrySdk;
 use Sentry\State\HubInterface;
 use function Sentry\init;
