+++ app/Factory/ContainerFactory.php
@@ -9,7 +9,7 @@ use CultuurNet\UDB3\JwtProvider\ApiGuardServiceProvider;
 use CultuurNet\UDB3\JwtProvider\Error\SentryWebServiceProvider;
 use CultuurNet\UDB3\JwtProvider\MiddlewareServiceProvider;
 use CultuurNet\UDB3\JwtProvider\RoutingServiceProvider;
-use CultuurNet\UDB3\JwtProvider\SentryHubServiceProvider;
+use CultuurNet\UDB3\JwtProvider\Error\SentryHubServiceProvider;
 use CultuurNet\UDB3\JwtProvider\SessionServiceProvider;
 use League\Container\Container;
 use League\Container\ReflectionContainer;
